requirements.txt
pyspark==3.5.1
delta-spark==3.1.0
pandas 
mlflow
fastapi
notebook==7.2.1
numpy
scikit-learn
evidently
matplotlib==3.9.1
python-box==6.0.2
pyYAML==6.0.1
ensure==1.0.2
xgboost
omegaconf==2.3.0
types-PyYAML==6.0.12.20240311
-e .